GED-api/app/api/middleware.py
import time
import uuid
from fastapi import Request
from starlette.middleware.base import BaseHTTPMiddleware
from app.core.database import AsyncSessionLocal
from app.models.audit import AuditLog
from jose import jwt
from app.core.config import settings

class AuditMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        # We can extract user details from request state or token for audit
        # This is a simplified middleware, capturing request actions.
        response = await call_next(request)
        
        # Only log specific altering methods or paths if needed
        # In this implementation, we log inside the business logic for specific document logic, 
        # but we can capture generic API paths here if wanted.
        return response
