GED-api/pyproject.toml
[project]
name = "ged"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=23.2.1",
    "alembic>=1.12.1",
    "asyncpg>=0.28.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.104.1",
    "minio>=7.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-jose[cryptography]>=3.3.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.6",
    "redis>=5.0.1",
    "sqlalchemy[asyncio]>=2.0.23",
    "uvicorn[standard]>=0.23.2",
]
